requirements.txt
aldryn-boilerplates==0.7.5
aldryn-bootstrap3==1.2.1
aldryn-forms==2.2.0
aldryn-google-analytics==1.0.1
aldryn-snake==0.0.2
appdirs==1.4.3
cmsplugin-filer==1.1.3
cssselect==1.0.1
cssutils==1.0.2
dj-database-url==0.4.2
Django==1.10.7
django-absolute==0.3
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.3
django-crispy-forms==1.6.1
django-durationfield==0.5.4
django-emailit==0.2.2
django-filer==1.2.7
django-formtools==2.0
django-guardian==1.4.6
django-haystack==2.6.1
django-machina==0.5.5
django-material==0.13.1
django-mptt==0.8.7
django-polymorphic==1.0.2
django-sekizai==0.10.0
django-simple-captcha==0.5.3
django-sizefield==0.9.1
django-tablib==3.1.2
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.2.0
djangocms-column==1.7.0
djangocms-googlemap==1.0.2
djangocms-installer==0.9.5
djangocms-link==2.1.1
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.4.0
djangocms-video==2.0.3
djangorestframework==3.6.2
easy-thumbnails==2.3
gunicorn==19.6.0
html5lib==0.9999999
lxml==3.7.3
markdown2==2.3.4
mysqlclient==1.3.10
olefile==0.44
packaging==16.8
Pillow==4.0.0
premailer==3.0.1
psycopg2==2.6.1
pyparsing==2.2.0
pytz==2016.10
qrcode==5.3
requests==2.13.0
scikit-learn==0.18.1
six==1.10.0
sklearn==0.0
tablib==0.11.4
tzlocal==1.3
Unidecode==0.4.20
whitenoise==2.0.6
YURL==0.13
